+++ walk.go
@@ -1,0 +1,32 @@
+package conditions
+
+// Visitor can be called by Walk to traverse an AST hierarchy.
+// The Visit() function is called once per node.
+type Visitor interface {
+	Visit(Node) Visitor
+}
+
+// Walk traverses a node hierarchy in depth-first order.
+func Walk(v Visitor, node Node) {
+	if v = v.Visit(node); v == nil {
+		return
+	}
+
+	switch n := node.(type) {
+	case *BinaryExpr:
+		Walk(v, n.LHS)
+		Walk(v, n.RHS)
+
+	case *ParenExpr:
+		Walk(v, n.Expr)
+	}
+}
+
+// WalkFunc traverses a node hierarchy in depth-first order.
+func WalkFunc(node Node, fn func(Node)) {
+	Walk(walkFuncVisitor(fn), node)
+}
+
+type walkFuncVisitor func(Node)
+
+func (fn walkFuncVisitor) Visit(n Node) Visitor { fn(n); return fn }
